+++ client/src/Pages/Collection/Tracks/Tracks.scss
@@ -6,22 +6,6 @@
     &__header{
         background-color: $purple;
         background-image: $grayLinear;
-        
-        button{
-            pointer-events: none;
-            width: 14.5rem;
-            height: 14.5rem;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            box-shadow: $cover;
-
-            .cover{
-                box-shadow: none;
-                width: 4.688rem;
-                height: 4.688rem;
-            }
-        }
     }
     
     &__main{
